services/service/requirements.txt
alembic==1.0.8
click==7.0
flask-cors==3.0.7
flask-migrate==1.8.0
flask-script==2.0.6
flask-sqlalchemy==2.1
flask==1.0.2
itsdangerous==1.1.0
jinja2==2.10
mako==1.0.8
markupsafe==1.1.1
numpy==1.16.2
pandas==0.24.2
psycopg2
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
six==1.12.0
sqlalchemy==1.3.1
werkzeug==0.15.1
